Fix: #51 `LeafTagRepository` 타입 수정
- update 반환 타입 int > Long 으로 변경

+++ src/main/java/Recorders/ggogit/domain/leaf/repository/LeafTagRepository.java
@@ -10,5 +10,5 @@ public interface LeafTagRepository {
     Long save(LeafTag leafTag);
     LeafTag findById(Long id);
     List<LeafTag> findAll();
-    int update(LeafTag leafTag);
+    Long update(LeafTag leafTag);
 }
